neki

+++ Kino/Kino/Models/Dvorana.cs
@@ -5,11 +5,11 @@ namespace Kino.Models
 {
     public partial class Dvorana
     {
-        public decimal IdDvorane { get; set; }
-        public decimal IdKolosej { get; set; }
-        public decimal IdTehnologije { get; set; }
+        public int IdDvorane { get; set; }
+        public int IdKolosej { get; set; }
+        public int IdTehnologije { get; set; }
         public string Ime { get; set; }
-        public decimal? StVrst { get; set; }
-        public decimal? StSedezovNaVrsto { get; set; }
+        public int? StVrst { get; set; }
+        public int? StSedezovNaVrsto { get; set; }
     }
 }
