+++ Kino/Kino/Models/Film.cs
@@ -5,10 +5,10 @@ namespace Kino.Models
 {
     public partial class Film
     {
-        public decimal IdFilm { get; set; }
-        public decimal IdZalozba { get; set; }
-        public decimal IdOmejitve { get; set; }
-        public decimal IdOsebjeFilma { get; set; }
+        public int IdFilm { get; set; }
+        public int IdZalozba { get; set; }
+        public int IdOmejitve { get; set; }
+        public int IdOsebjeFilma { get; set; }
         public string Naslov { get; set; }
         public DateTime? Leto { get; set; }
         public DateTime? CasTrajanja { get; set; }
