+++ Kino/Kino/Models/IgraV.cs
@@ -5,7 +5,7 @@ namespace Kino.Models
 {
     public partial class IgraV
     {
-        public decimal IdFilm { get; set; }
-        public decimal IdOsebjeFilma { get; set; }
+        public int IdFilm { get; set; }
+        public int IdOsebjeFilma { get; set; }
     }
 }
