+++ Kino/Kino/Models/Karta.cs
@@ -5,9 +5,9 @@ namespace Kino.Models
 {
     public partial class Karta
     {
-        public decimal IdSedeza { get; set; }
-        public decimal IdFilm { get; set; }
-        public decimal IdDvorane { get; set; }
+        public int IdSedeza { get; set; }
+        public int IdFilm { get; set; }
+        public int IdDvorane { get; set; }
         public DateTime CasZacetka { get; set; }
         public DateTime CasKonca { get; set; }
         public DateTime Datum { get; set; }
