+++ Kino/Kino/Models/OsebjeFilma.cs
@@ -5,7 +5,7 @@ namespace Kino.Models
 {
     public partial class OsebjeFilma
     {
-        public decimal IdOsebjeFilma { get; set; }
+        public int IdOsebjeFilma { get; set; }
         public string Ime { get; set; }
         public DateTime? DatumRojstva { get; set; }
     }
