+++ Kino/Kino/Models/Pozicija.cs
@@ -5,7 +5,7 @@ namespace Kino.Models
 {
     public partial class Pozicija
     {
-        public decimal IdPozicije { get; set; }
+        public int IdPozicije { get; set; }
         public string ImePozicije { get; set; }
     }
 }
