+++ Kino/Kino/Models/Predstava.cs
@@ -5,8 +5,8 @@ namespace Kino.Models
 {
     public partial class Predstava
     {
-        public decimal IdFilm { get; set; }
-        public decimal IdDvorane { get; set; }
+        public int IdFilm { get; set; }
+        public int IdDvorane { get; set; }
         public DateTime CasZacetka { get; set; }
         public DateTime CasKonca { get; set; }
         public DateTime Datum { get; set; }
