+++ Kino/Kino/Models/Pripada.cs
@@ -5,7 +5,7 @@ namespace Kino.Models
 {
     public partial class Pripada
     {
-        public decimal IdZvrst { get; set; }
-        public decimal IdFilm { get; set; }
+        public int IdZvrst { get; set; }
+        public int IdFilm { get; set; }
     }
 }
