+++ Kino/Kino/Models/StarostnaOmejitev.cs
@@ -5,7 +5,7 @@ namespace Kino.Models
 {
     public partial class StarostnaOmejitev
     {
-        public decimal IdOmejitve { get; set; }
+        public int IdOmejitve { get; set; }
         public string Ime { get; set; }
     }
 }
